requirements.txt
pandas>=2.0
python-dotenv>=1.0
langchain-openai>=0.1
requests>=2.31